Fix server status displayed as unknown
When the api endpoint of the management server returned the 'UP'
state, the status was displayed as unknown, because the greeter
internally uses 'ACTIVE' instead of 'UP'. The solution is to set
the detected state to 'ACTIVE' when the server responds with 'UP'.
Additionally the response code check is no longer necessary.